+++ src/Baldr/Baldr.hpp
@@ -23,6 +23,7 @@
 #include <Baldr/Middleware/SecurityHeaders.hpp>
 #include <Baldr/OpenApi/BaldrOpenApiExtension.hpp>
 #include <Baldr/OpenApi/MapOpenApi.hpp>
+#include <Baldr/OpenApi/MapScalarUi.hpp>
 namespace BALDR_NAMESPACE
 {